requirements.txt
fastapi
uvicorn
llama-cpp-python
pydantic
sentence-transformers
chromadb
pdfplumber